RoundRobinLoadBalancer adjust selection and connection storage
Motivation:
RoundRobinLoadBalancer's Host class maintains a list of active
connections in a copy-on-write fashion. However when an item is added or
removed their is an additional copy operation due to initial state not
accounting for 1 more/less element. Also the connection selection
algorithm relies upon random selection. With less than 64 connections
there are as many attempts as connections to select, but no accounting
for overlapping indexes returned from random. It is highly likely a
duplicate selection attempt in the same loop with fail on repeat
attempts. More failed attempts means more atomic operations and more
likely to create new connections when not necessary. When the number of
connections is gteq 64 many attempts maybe made to select a connection.
Since we are doing random index selection we can attempt a fixed amount
and create a new connection if we fail to find an available connection.

Modifications:
- Host class uses an array for copy-on-write storage and allocates
  exactly the required amount of space on add/remove.
- Connection selection algorithm avoids duplicates when number
  connections <= 64, otherwise attempts fixed random selection attempts.

Result:
Less resize/copy operations on connection established/closed, less
duplicate connection selection attempts.