Avoid error from connection closure when signals should be queued
Motivation:
The HTTP protocol knows the request/response processing is over in some
scenarios before the read or write Publisher terminates. For example if
content-length is 0 we know there is no payload, and if content-length
is present for HTTP/1.x we know there will be no trailers. Internally we
use an EmptyTrailers object as a marker to indicate "end of Publisher",
but this may result in premature errors in the event the protocol knows
we are done before internal state is updated, and socket close events
occur. This problem is exacerbated with offloading enabled and events
are processed on different threads.

Modifications:
- Remove the requirement that an EmptyTrailers object is always present
  to signify "end of Publisher". CloseHandler knows when the stream has
  ended and generates an event that ensures NettyChannelPublisher state
  has been updated.
- Increase robustness of HttpObjectEncoder to verify if content-length
  is present that the exact size must be written, and no non-empty
  trailers are allowed.

Result:
Fixes https://github.com/apple/servicetalk/issues/1507
Fixes https://github.com/apple/servicetalk/issues/1579